+++ recipes/config_render.rb
@@ -17,9 +17,9 @@
 # set the config path based on default attributes
 # render out our config
 zookeeper_config 'zookeeper config' do
-  path   "#{node[:zookeeper][:config_dir] % { zookeeper_version: node[:zookeeper][:version] }}/" \
-         "#{node[:zookeeper][:conf_file]}"
-  config node[:zookeeper][:config]
-  user   node[:zookeeper][:user]
+  path   "#{node['zookeeper']['config_dir'] % { zookeeper_version: node['zookeeper']['version'] }}/" \
+         "#{node['zookeeper']['conf_file']}"
+  config node['zookeeper']['config']
+  user   node['zookeeper']['user']
   action :render
 end