+++ recipes/install.rb
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-if node[:zookeeper][:use_java_cookbook] == true
+if node['zookeeper']['use_java_cookbook'] == true
   include_recipe 'java::default'
 else
   Chef::Log.info("Assuming you've provided your own Java")
@@ -26,21 +26,21 @@ include_recipe 'apt::default' if node['platform_family'] == 'debian'
 node.override['build-essential']['compile_time'] = true
 include_recipe 'build-essential::default'
 
-zookeeper node[:zookeeper][:version] do
-  user        node[:zookeeper][:user]
-  user_home   node[:zookeeper][:user_home]
-  mirror      node[:zookeeper][:mirror]
-  checksum    node[:zookeeper][:checksum]
-  install_dir node[:zookeeper][:install_dir]
-  log_dir     node[:zookeeper][:log_dir]
-  data_dir    node[:zookeeper][:config][:dataDir]
+zookeeper node['zookeeper']['version'] do
+  user        node['zookeeper']['user']
+  user_home   node['zookeeper']['user_home']
+  mirror      node['zookeeper']['mirror']
+  checksum    node['zookeeper']['checksum']
+  install_dir node['zookeeper']['install_dir']
+  log_dir     node['zookeeper']['log_dir']
+  data_dir    node['zookeeper']['config']['dataDir']
   action      :install
 end
 
 # Add optional Zookeeper environment vars
-if node[:zookeeper][:env_vars]
-  file "#{node[:zookeeper][:config_dir]}/zookeeper-env.sh" do
-    owner node[:zookeeper][:user]
-    content exports_config node[:zookeeper][:env_vars]
+if node['zookeeper']['env_vars']
+  file "#{node['zookeeper']['config_dir']}/zookeeper-env.sh" do
+    owner node['zookeeper']['user']
+    content exports_config node['zookeeper']['env_vars']
   end
 end