+++ test/integration/sysv/serverspec/sysv_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-context 'when node[:zookeeper][:service_style] is set to sysv on CentOS' do
+context 'when node[zookeeper][service_style] is set to sysv on CentOS' do
   describe file '/etc/init.d/zookeeper' do
     it { is_expected.to be_file }
   end