+++ CHANGELOG.md
@@ -61,7 +61,7 @@ This file is used to list changes made in each version of zookeeper.
 ## v2.11.0
 
 * Fix logic around creating `zookeeper-env.sh` (Fixes #141)
-* Add tests for default attribtues & using `node[:zookeeper][:env_vars]`
+* Add tests for default attributes & using `node[zookeeper][env_vars]`
 * Add JAVA_OPTS attribute (#144, thanks @andrewgoktepe)
 
 ## v2.10.0